SEQUOIACM-1094:重构文件元数据处理流程

+++ src/content-server/project/common/content-module/src/main/java/com/sequoiacm/contentserver/dao/ScmFileDataDeleterWrapper.java
@@ -1,13 +1,9 @@
 package com.sequoiacm.contentserver.dao;
 
 import com.sequoiacm.common.CommonHelper;
-import com.sequoiacm.common.FieldName;
 import com.sequoiacm.common.ScmFileLocation;
 import com.sequoiacm.contentserver.model.ScmWorkspaceInfo;
-import com.sequoiacm.contentserver.site.ScmContentModule;
-import com.sequoiacm.datasource.dataoperation.ScmDataInfo;
-import com.sequoiacm.infrastructure.common.BsonUtils;
-import org.bson.BSONObject;
+import com.sequoiacm.contentserver.pipeline.file.module.FileMeta;
 import org.bson.types.BasicBSONList;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,31 +14,34 @@ import java.util.List;
 public class ScmFileDataDeleterWrapper {
     private static final Logger logger = LoggerFactory.getLogger(ScmFileDataDeleterWrapper.class);
     private final ScmWorkspaceInfo wsInfo;
-    private final BSONObject file;
+    private FileMeta fileMeta;
 
-    public ScmFileDataDeleterWrapper(ScmWorkspaceInfo wsInfo, BSONObject file) {
-        this.file = file;
+
+    public ScmFileDataDeleterWrapper(ScmWorkspaceInfo wsInfo, FileMeta fileMeta) {
+        this.fileMeta = fileMeta;
         this.wsInfo = wsInfo;
     }
 
     public void deleteDataSilence() {
-        ScmDataInfo dataInfo = new ScmDataInfo(file);
+        if (fileMeta.isDeleteMarker()) {
+            return;
+        }
         try {
-            BasicBSONList sites = (BasicBSONList) file.get(FieldName.FIELD_CLFILE_FILE_SITE_LIST);
+            BasicBSONList sites = fileMeta.getSiteList();
             List<ScmFileLocation> siteList = new ArrayList<>();
             CommonHelper.getFileLocationList(sites, siteList);
             List<Integer> siteIdList = new ArrayList<>();
             for (ScmFileLocation fileLocation : siteList) {
                 siteIdList.add(fileLocation.getSiteId());
             }
-            ScmFileDataDeleter dataDeleter = new ScmFileDataDeleter(siteIdList, wsInfo, dataInfo);
+            ScmFileDataDeleter dataDeleter = new ScmFileDataDeleter(siteIdList, wsInfo,
+                    fileMeta.getDataInfo());
             dataDeleter.deleteData();
         }
         catch (Exception e) {
             logger.warn("remove file data failed:fileId={},version={}.{},dataInfo={}",
-                    BsonUtils.getStringChecked(file, FieldName.FIELD_CLFILE_ID),
-                    file.get(FieldName.FIELD_CLFILE_MAJOR_VERSION),
-                    file.get(FieldName.FIELD_CLFILE_MINOR_VERSION), dataInfo, e);
+                    fileMeta.getId(), fileMeta.getMajorVersion(), fileMeta.getMinorVersion(),
+                    fileMeta.getDataInfo(), e);
         }
     }
 }
